fix #12314 - Validator: do not attempt to delete twice
git-svn-id: http://josm.openstreetmap.de/svn/trunk@9359 0c6e7542-c601-0410-84e7-c038aed88b3b

+++ src/org/openstreetmap/josm/data/validation/tests/MapCSSTagChecker.java
@@ -504,7 +504,7 @@ public class MapCSSTagChecker extends Test.TagTest {
             for (FixCommand fixCommand : fixCommands) {
                 cmds.add(fixCommand.createCommand(p, matchingSelector));
             }
-            if (deletion) {
+            if (deletion && !p.isDeleted()) {
                 cmds.add(new DeleteCommand(p));
             }
             return new SequenceCommand(tr("Fix of {0}", getDescriptionForMatchingSelector(p, matchingSelector)), cmds);
